Working commit - progress in data generation script + s3 seed finalized
