+++ database/seed/s3.seed.js
@@ -10,8 +10,7 @@ const stream = require('stream');
 const Promise = require('promise');
 const AWS = require('./aws.js');
 
-const creds = new AWS.SharedIniFileCredentials({ profile: 'default' });
-AWS.config.credentials = creds;
+AWS.config.credentials = new AWS.SharedIniFileCredentials({ profile: 'default' });
 
 const httpsAgent = new https.Agent({
   keepAlive: true,
@@ -36,6 +35,7 @@ const streamToS3 = async (imgNum, width, height) => {
       Key: `${s3KeyBase}/img${imgNum}.jpeg`,
       ContentType: contentType,
     };
+
     s3promise = s3.upload(params).promise();
     return passThrough;
   };
@@ -66,7 +66,7 @@ let randomWidth = null;
 let randomHeight = null;
 const s3PromiseArray = [];
 
-for (let i = 1; i <= 1; i += 1) {
+for (let i = 1; i <= 1000; i += 1) {
   // generate sizeOptions for picture requst
   randomWidth = sizeOptions[Math.floor(Math.random() * 2)];
   randomHeight = sizeOptions[Math.floor(Math.random() * 2)];
@@ -74,6 +74,7 @@ for (let i = 1; i <= 1; i += 1) {
 }
 
 // Resolve picture upload stream promises
+// Note -- this will return an array of nulls due to scoping & async operations.
 // eslint-disable-next-line no-console
 Promise.all(s3PromiseArray).then((results) => console.log('promise all results --->', results))
 // eslint-disable-next-line no-console
